feat(syntax,transformer): move JSX entity decoder to `oxc_syntax` (#24617)
## Summary

- move the JSX entity decoder from `oxc_transformer` to `oxc_syntax`
- keep the existing accumulator-based API and transformer call sites unchanged
- move the stray-ampersand regression test with the decoder

This is a behavior-neutral preparation for handling JSX entity decoding closer to parsing in a follow-up.
